+++ account/utils.py
@@ -125,13 +125,19 @@ def check_password_expired(user):
         # use global value
         expiry = settings.ACCOUNT_PASSWORD_EXPIRY
 
+    if expiry == 0:  # zero indicates no expiration
+        return False
+
     try:
         # get latest password info
         latest = user.password_history.latest("timestamp")
     except PasswordHistory.DoesNotExist:
         return False
 
-    if datetime.datetime.now(tz=pytz.UTC) > (latest.timestamp + datetime.timedelta(seconds=expiry)):
-        return False
+    now = datetime.datetime.now(tz=pytz.UTC)
+    expiration = latest.timestamp + datetime.timedelta(seconds=expiry)
 
-    return True
+    if expiration < now:
+        return True
+    else:
+        return False