+++ docs/changelog.rst
@@ -3,6 +3,13 @@
 CHANGELOG
 =========
 
+2.0
+---
+
+* add password expiration
+* add Django v1.10 to test compatibility matrix
+* remove Python 3.2 from text compatibility matrix
+
 1.0
 ---